+++ CHANGELOG.md
@@ -19,6 +19,7 @@ handle multi-line list body as single line.
 
 - Improve category detection for TV and Movies in Torrents-CSV search provider
 - Update YTS provider domain proxy to .bz
+- Update YTS provider API URL
 
 ### Removed